feat (connection): don't cache API_ROOT at runtime
The only way ParsePy provides to use a different API_ROOT is by setting an
environment variable before importing `parse_rest`, which is awkward.
After import, API_ROOT is embedded in endpoints cached on a number of
classes, making it difficult to change.  With this hack, all those
endpoints will be generated dynamically as needed, so API_ROOT can be
changed after import.